Safari-611.1.10/LayoutTests/media/modern-media-controls/seek-backward-support/seek-backward-support-expected.txt

Testing the rewind button in fullscreen.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS !!window.internals.shadowRoot(media).lastChild.querySelector("button.rewind") became true
PASS media.webkitDisplayingFullscreen is true

Setting media.currentTime to 5s
PASS media.currentTime is 5
PASS media.paused is true

Starting to press on the skip back button
PASS media.currentTime < 4.5 became true
PASS media.paused is true

PASS successfullyParsed is true

TEST COMPLETE
